test(core): migrate UpdateThemeTests (Wave C)
14 tests, 26 failure sites preserved 1:1. Three `guard let restored = validateAndBuild` single-value unwraps convert to `try #require`; the one .exportState case destructure to Issue.record + return.
