--- tests/UpdateThemeTests.swift
@@ -1,202 +1,0 @@
-// Tests for per-pane theme switching: setPaneTheme message, split inheritance,
-// snapshot round-tripping, and resilient decoding of unknown theme names.
-import Foundation
-
-func themeTests() {
-    print("Theme Tests...")
-
-    test("setPaneTheme updates model") {
-        var model = makeModel()
-        createTab(&model)
-        let paneId = model.groups[0].tabs[0].focusedPaneId
-        update(&model, .setPaneTheme(paneId: paneId, themeName: "Dracula"))
-        try expectEqual(model.pane(paneId)?.theme, "Dracula")
-    }
-
-    test("setPaneTheme projects theme and schedules checkpoint") {
-        var model = makeModel()
-        createTab(&model)
-        let paneId = model.groups[0].tabs[0].focusedPaneId
-        let commands = update(&model, .setPaneTheme(paneId: paneId, themeName: "Nord"))
-        try expectEqual(desiredPaneConfig(in: model)[paneId]?.theme, "Nord")
-        try expect(hasEffect(commands) {
-            if case .scheduleCheckpoint = $0 { return true }; return false
-        }, "should emit scheduleCheckpoint")
-    }
-
-    test("clearPaneTheme sets theme to nil") {
-        var model = makeModel()
-        createTab(&model)
-        let paneId = model.groups[0].tabs[0].focusedPaneId
-        update(&model, .setPaneTheme(paneId: paneId, themeName: "Dracula"))
-        try expectEqual(model.pane(paneId)?.theme, "Dracula")
-        update(&model, .setPaneTheme(paneId: paneId, themeName: nil))
-        try expect(model.pane(paneId)?.theme == nil, "theme should be nil after clearing")
-    }
-
-    test("splitPane inherits theme from parent") {
-        var model = makeModel()
-        createTab(&model)
-        let paneId = model.groups[0].tabs[0].focusedPaneId
-        update(&model, .setPaneTheme(paneId: paneId, themeName: "Catppuccin Mocha"))
-        update(&model, .splitPane(paneId: paneId, direction: .horizontal))
-        let tab = selectedTab(in: model)!
-        let newPaneId = tab.focusedPaneId
-        try expect(newPaneId != paneId, "new pane should be different from parent")
-        try expectEqual(model.pane(newPaneId)?.theme, "Catppuccin Mocha")
-    }
-
-    test("splitPane with theme projects config for new pane") {
-        var model = makeModel()
-        createTab(&model)
-        let paneId = model.groups[0].tabs[0].focusedPaneId
-        update(&model, .setPaneTheme(paneId: paneId, themeName: "Rose Pine"))
-        update(&model, .splitPane(paneId: paneId, direction: .vertical))
-        let tab = selectedTab(in: model)!
-        let newPaneId = tab.focusedPaneId
-        try expectEqual(desiredPaneConfig(in: model)[newPaneId]?.theme, "Rose Pine")
-    }
-
-    test("splitPane without theme has no projected config") {
-        var model = makeModel()
-        createTab(&model)
-        update(&model, .splitPane(direction: .horizontal))
-        let paneId = selectedTab(in: model)!.focusedPaneId
-        try expect(desiredPaneConfig(in: model)[paneId] == nil, "no theme -> no config key")
-    }
-
-    test("toSnapshot preserves theme") {
-        var model = makeModel()
-        createTab(&model)
-        let paneId = model.groups[0].tabs[0].focusedPaneId
-        update(&model, .setPaneTheme(paneId: paneId, themeName: "Gruvbox Dark"))
-        let snapshot = toSnapshot(model)
-        let ps = paneSnapshot(paneId.rawValue.uuidString, in: snapshot)
-        try expectEqual(ps?.theme, "Gruvbox Dark")
-    }
-
-    test("snapshot round-trip preserves theme") {
-        var model = makeModel()
-        createTab(&model)
-        let paneId = model.groups[0].tabs[0].focusedPaneId
-        update(&model, .setPaneTheme(paneId: paneId, themeName: "TokyoNight Night"))
-        let snapshot = toSnapshot(model)
-        guard let restored = validateAndBuild(snapshot) else {
-            throw TestFailure(message: "snapshot round-trip failed")
-        }
-        let restoredPane = restored.pane(paneId)
-        try expect(restoredPane != nil, "pane should exist in restored model")
-        try expectEqual(restoredPane?.theme, "TokyoNight Night")
-    }
-
-    test("snapshot with unknown theme name decodes and preserves in model") {
-        let json = """
-        {
-          "version": 2,
-          "model": {
-            "groups": [{
-              "id": "E53A57E9-1B39-4E15-B2AD-CA6B8700F17A",
-              "name": "General",
-              "tabs": [{
-                "id": "89B4C232-C840-42A8-8CA6-C133C8EBBFF2",
-                "rootNode": { "type": "leaf", "pane": {
-                  "id": "A13076E4-A29C-4358-A771-B4B4DF84C6C5",
-                  "title": "Terminal",
-                  "theme": "NonExistent Theme"
-                } }
-              }]
-            }]
-          }
-        }
-        """
-        let data = json.data(using: .utf8)!
-        let initFile = try JSONDecoder().decode(AppInitFile.self, from: data)
-        try expectEqual(allPaneSnapshots(initFile.model)[0].theme, "NonExistent Theme")
-        // validateAndBuild should succeed — unknown themes are preserved as-is
-        let built = validateAndBuild(initFile.model)
-        try expect(built != nil, "should rebuild despite unknown theme")
-        let paneId = PaneId(rawValue: UUID(uuidString: "A13076E4-A29C-4358-A771-B4B4DF84C6C5")!)
-        try expectEqual(built!.pane(paneId)?.theme, "NonExistent Theme")
-    }
-
-    test("export preserves theme in snapshot") {
-        var model = makeModel()
-        createTab(&model)
-        let paneId = model.groups[0].tabs[0].focusedPaneId
-        update(&model, .setPaneTheme(paneId: paneId, themeName: "Catppuccin Latte"))
-        let commands = update(&model, .exportState)
-        guard case .exportState(let snapshot) = commands.first else {
-            throw TestFailure(message: "expected exportState command")
-        }
-        let ps = paneSnapshot(paneId.rawValue.uuidString, in: snapshot)
-        try expectEqual(ps?.theme, "Catppuccin Latte")
-    }
-
-    test("mergeCheckpoints preserves theme") {
-        let p1 = PaneId()
-        let light = ValidatedAppRestore(
-            snapshot: toSnapshot(makeModel()), model: makeModel(),
-            paneSnapshots: [p1: PaneSnapshot(id: p1.rawValue.uuidString, title: "t", cwd: "/c", launch: nil, scrollback: nil, theme: "Dracula")]
-        )
-        let enriched = ValidatedAppRestore(
-            snapshot: toSnapshot(makeModel()), model: makeModel(),
-            paneSnapshots: [p1: PaneSnapshot(id: p1.rawValue.uuidString, title: "t", cwd: "/c", launch: nil, scrollback: "text", theme: "Dracula")]
-        )
-        let merged = mergeCheckpoints(light: light, enriched: enriched)
-        try expectEqual(merged.paneSnapshots[p1]?.theme, "Dracula")
-        try expectEqual(merged.paneSnapshots[p1]?.scrollback, "text")
-    }
-
-    test("theme name round-trips as arbitrary string") {
-        var model = makeModel()
-        createTab(&model)
-        let paneId = model.groups[0].tabs[0].focusedPaneId
-        update(&model, .setPaneTheme(paneId: paneId, themeName: "My Custom Theme"))
-        let snapshot = toSnapshot(model)
-        guard let restored = validateAndBuild(snapshot) else {
-            throw TestFailure(message: "snapshot round-trip failed")
-        }
-        try expectEqual(restored.pane(paneId)?.theme, "My Custom Theme")
-    }
-
-    test("unknown theme name preserved in snapshot round-trip") {
-        let json = """
-        {
-          "version": 2,
-          "model": {
-            "groups": [{
-              "id": "E53A57E9-1B39-4E15-B2AD-CA6B8700F17A",
-              "name": "General",
-              "tabs": [{
-                "id": "89B4C232-C840-42A8-8CA6-C133C8EBBFF2",
-                "rootNode": { "type": "leaf", "pane": {
-                  "id": "A13076E4-A29C-4358-A771-B4B4DF84C6C5",
-                  "title": "Terminal",
-                  "theme": "NonExistent Theme"
-                } }
-              }]
-            }]
-          }
-        }
-        """
-        let data = json.data(using: .utf8)!
-        let initFile = try JSONDecoder().decode(AppInitFile.self, from: data)
-        let built = validateAndBuild(initFile.model)
-        try expect(built != nil, "should rebuild")
-        // Round-trip the built model back to snapshot
-        let snapshot = toSnapshot(built!)
-        try expectEqual(allPaneSnapshots(snapshot)[0].theme, "NonExistent Theme")
-    }
-
-    test("nil theme round-trips through snapshot") {
-        var model = makeModel()
-        createTab(&model)
-        let paneId = model.groups[0].tabs[0].focusedPaneId
-        // Don't set any theme — it's nil by default
-        let snapshot = toSnapshot(model)
-        guard let restored = validateAndBuild(snapshot) else {
-            throw TestFailure(message: "snapshot round-trip failed")
-        }
-        try expect(restored.pane(paneId)?.theme == nil, "nil theme should survive round-trip")
-    }
-}
